requirements.txt
# https://pyyaml.org/wiki/PyYAMLDocumentation
pyyaml